requirements.txt
seccs>=0.0.5
coverage>=4.1
pytest-cov>=2.2.1
mock>=2.0.0
argh>=0.26.2
tabulate>=0.8.0
msgpack-python>=0.4.8
pyyaml>=3.12
flask>=0.12.2
requests>=2.18.4
